refactor(product-image): move product-image and all it's sister components to their own directory within the root components directory; implement changes to imports accordingly;
