+++ src/app/collections/[collection-slug]/[product-slug]/components/product-provider.tsx
@@ -1,4 +1,6 @@
-import { ProductImageGroupProvider } from "./product-image-group-context";
+"use client"
+
+import { ProductImageGroupProvider } from "@/components/product-image";
 
 export function ProductProvider({ children }: React.PropsWithChildren) {
     return <ProductImageGroupProvider>{children}</ProductImageGroupProvider>;
