+++ src/components/product-image/index.ts
@@ -1,0 +1,3 @@
+export * from "./product-image";
+export * from "./product-image-group";
+export * from "./product-image-group-context";
