+++ src/components/product-image/product-image-group-context.tsx
