+++ src/components/product-image/product-image.tsx
